Move attribute utils into a separate util class
