+++ plugin/src/main/java/com/denizenscript/denizen/objects/properties/entity/EntityAttributeModifiers.java
@@ -1,9 +1,7 @@
 package com.denizenscript.denizen.objects.properties.entity;
 
-import com.denizenscript.denizen.nms.NMSHandler;
-import com.denizenscript.denizen.nms.NMSVersion;
 import com.denizenscript.denizen.objects.EntityTag;
-import com.denizenscript.denizen.objects.properties.item.ItemAttributeModifiers;
+import com.denizenscript.denizen.utilities.AttributeUtil;
 import com.denizenscript.denizen.utilities.BukkitImplDeprecations;
 import com.denizenscript.denizen.utilities.Utilities;
 import com.denizenscript.denizencore.objects.Mechanism;
@@ -13,7 +11,6 @@ import com.denizenscript.denizencore.objects.core.ListTag;
 import com.denizenscript.denizencore.objects.core.MapTag;
 import com.denizenscript.denizencore.objects.properties.Property;
 import com.denizenscript.denizencore.objects.properties.PropertyParser;
-import com.denizenscript.denizencore.tags.TagContext;
 import com.denizenscript.denizencore.tags.core.EscapeTagUtil;
 import com.denizenscript.denizencore.utilities.CoreUtilities;
 import com.denizenscript.denizencore.utilities.debugging.Debug;
@@ -23,7 +20,6 @@ import org.bukkit.attribute.Attribute;
 import org.bukkit.attribute.AttributeInstance;
 import org.bukkit.attribute.AttributeModifier;
 import org.bukkit.inventory.EquipmentSlot;
-import org.bukkit.inventory.EquipmentSlotGroup;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -57,12 +53,6 @@ public class EntityAttributeModifiers implements Property {
     EntityTag entity;
 
     @Deprecated
-    public static String stringify(AttributeModifier modifier) {
-        return EscapeTagUtil.escape(modifier.getName()) + "/" + modifier.getAmount() + "/" + modifier.getOperation().name()
-                + "/" + (modifier.getSlot() == null ? "any" : modifier.getSlot().name());
-    }
-
-    @Deprecated
     public ListTag getAttributes() {
         ListTag list = new ListTag();
         for (Attribute attribute : Attribute.values()) {
@@ -72,96 +62,11 @@ public class EntityAttributeModifiers implements Property {
             }
             StringBuilder modifiers = new StringBuilder();
             for (AttributeModifier modifier : instance.getModifiers()) {
-                modifiers.append("/").append(stringify(modifier));
+                modifiers.append("/").append(AttributeUtil.modifierToLegacyString(modifier));
             }
-            list.add(EscapeTagUtil.escape(ItemAttributeModifiers.legacyAttributeName(attribute)) + "/" + instance.getBaseValue() + modifiers);
+            list.add(EscapeTagUtil.escape(AttributeUtil.legacyName(attribute)) + "/" + instance.getBaseValue() + modifiers);
         }
         return list;
-    }
-
-    public static MapTag mapify(AttributeModifier modifier) {
-        MapTag result = new MapTag();
-        result.putObject("name", new ElementTag(modifier.getName()));
-        result.putObject("amount", new ElementTag(modifier.getAmount()));
-        result.putObject("operation", new ElementTag(modifier.getOperation()));
-        if (NMSHandler.getVersion().isAtLeast(NMSVersion.v1_20)) {
-            result.putObject("slot", new ElementTag(modifier.getSlotGroup().toString(), true));
-        }
-        else {
-            result.putObject("slot", new ElementTag(modifier.getSlot() == null ? "any" : modifier.getSlot().name()));
-        }
-        if (ItemAttributeModifiers.MODERN_ATTRIBUTE_FORMAT) {
-            result.putObject("key", new ElementTag(Utilities.namespacedKeyToString(modifier.getKey()), true));
-        }
-        // TODO: remove/deprecate the UUID key
-        result.putObject("id", new ElementTag(modifier.getUniqueId().toString()));
-        return result;
-    }
-
-    public static AttributeModifier modiferForMap(Attribute attr, MapTag map, TagContext context) {
-        ElementTag amount = map.getElement("amount");
-        ElementTag operation = map.getElement("operation");
-        double amountValue;
-        AttributeModifier.Operation operationValue = operation.asEnum(AttributeModifier.Operation.class);
-        if (operationValue == null) {
-            Debug.echoError("Attribute modifier operation '" + operation + "' does not exist.");
-            return null;
-        }
-        try {
-            amountValue = Double.parseDouble(amount.toString());
-        }
-        catch (NumberFormatException ex) {
-            Debug.echoError("Attribute modifier amount '" + amount + "' is not a valid decimal number.");
-            return null;
-        }
-        if (!NMSHandler.getVersion().isAtLeast(NMSVersion.v1_21)) {
-            return parseLegacyModifier(attr, map, amountValue, operationValue);
-        }
-        ElementTag key = map.getElement("key");
-        if (key == null && map.size() >= 2) {
-            BukkitImplDeprecations.pre1_21AttributeFormat.warn(context);
-            return parseLegacyModifier(attr, map, amountValue, operationValue);
-        }
-        if (key == null) {
-            Debug.echoError("Must specify a key.");
-            return null;
-        }
-        String slotGroupName = map.getElement("slot", "any").asString();
-        EquipmentSlotGroup group = EquipmentSlotGroup.getByName(slotGroupName);
-        if (group == null) {
-            EquipmentSlot slot = ElementTag.asEnum(EquipmentSlot.class, slotGroupName);
-            if (slot == null) {
-                Debug.echoError("Invalid equipment slot group specified: " + slotGroupName);
-                return null;
-            }
-            group = slot.getGroup();
-        }
-        return new AttributeModifier(Utilities.parseNamespacedKey(key.asString()), amountValue, operationValue, group);
-    }
-
-    @Deprecated(forRemoval = true)
-    public static AttributeModifier parseLegacyModifier(Attribute attr, MapTag map, double amount, AttributeModifier.Operation operation) {
-        ElementTag name = map.getElement("name");
-        ElementTag slot = map.getElement("slot", "any");
-        ElementTag id = map.getElement("id");
-        UUID idValue;
-        try {
-            idValue = id == null ? UUID.randomUUID() : UUID.fromString(id.toString());
-        }
-        catch (IllegalArgumentException ex) {
-            Debug.echoError("Attribute modifier ID '" + id + "' is not a valid UUID.");
-            return null;
-        }
-        EquipmentSlot slotValue = CoreUtilities.equalsIgnoreCase(slot.toString(), "any") ? null : slot.asEnum(EquipmentSlot.class);
-        if (slotValue == null && NMSHandler.getVersion().isAtLeast(NMSVersion.v1_20)) {
-            EquipmentSlotGroup group = EquipmentSlotGroup.getByName(slot.asString());
-            if (group == null) {
-                Debug.echoError("Invalid equipment slot group specified: " + slot);
-                return null;
-            }
-            return new AttributeModifier(idValue, name == null ? ItemAttributeModifiers.legacyAttributeName(attr) : name.asString(), amount, operation, group);
-        }
-        return new AttributeModifier(idValue, name == null ? ItemAttributeModifiers.legacyAttributeName(attr) : name.toString(), amount, operation, slotValue);
     }
 
     public ListTag getAttributeModifierList(AttributeInstance instance) {
@@ -172,7 +77,7 @@ public class EntityAttributeModifiers implements Property {
         if (modifiers.isEmpty()) {
             return null;
         }
-        return new ListTag(modifiers, EntityAttributeModifiers::mapify);
+        return new ListTag(modifiers, AttributeUtil::modifierToMap);
     }
 
     public MapTag getAttributeModifiers(boolean includeDeprecated) {
@@ -180,7 +85,7 @@ public class EntityAttributeModifiers implements Property {
         for (Attribute attribute : Utilities.listTypesRaw(Attribute.class)) {
             ListTag list = getAttributeModifierList(getAttributable().getAttribute(attribute));
             if (list != null) {
-                ItemAttributeModifiers.addAttributeToMap(map, attribute, list, includeDeprecated);
+                AttributeUtil.addToMap(map, attribute, list, includeDeprecated);
             }
         }
         return map;
@@ -200,72 +105,6 @@ public class EntityAttributeModifiers implements Property {
     public String getPropertyId() {
         return "attribute_modifiers";
     }
-
-
-    // <--[language]
-    // @name Attribute Modifiers
-    // @group Properties
-    // @description
-    // In minecraft, the "attributes" system defined certain core numerical values on entities, such as max health or attack damage.
-    // The value of an "attribute" is determined by its "base value" modified mathematically by each of its "attribute modififers".
-    // "Attribute modifiers" can be added either directly to the entity, or onto items - when on an item, an entity can equip it into the correct slot to automatically apply the modifier.
-    //
-    // These can be read via such tags as <@link tag EntityTag.attribute_modifiers>, <@link tag ItemTag.attribute_modifiers>,
-    // <@link tag EntityTag.has_attribute>, <@link tag EntityTag.attribute_value>, <@link tag EntityTag.attribute_base_value>, <@link tag EntityTag.attribute_default_value>, ...
-    //
-    // These can be modified by such mechanisms as <@link mechanism EntityTag.attribute_base_values>, <@link mechanism EntityTag.attribute_modifiers>, <@link mechanism EntityTag.add_attribute_modifiers>,
-    // <@link mechanism EntityTag.remove_attribute_modifiers>, <@link mechanism ItemTag.attribute_modifiers>, <@link mechanism ItemTag.add_attribute_modifiers>, <@link mechanism ItemTag.remove_attribute_modifiers>, ...
-    //
-    // The input format of each of the 'add' and set mechanisms is slightly complicated: a MapTag where the keys are attribute names, and values are a ListTag of modifiers,
-    // where each modifier is itself a MapTag with required keys 'operation' and 'amount', and additionally:
-    // Before MC 1.21: optional 'name', 'slot', and 'id' keys.
-    // The default ID will be randomly generated, the default name will be the attribute name.
-    // After MC 1.21: required 'key' key, and optional 'slot'.
-    // The 'key' is the attribute's name/identifier in a "namespace:key" format (defaulting to the "minecraft" namespace), which has to be distinct to other modifiers of the same type on the object.
-    //
-    // Valid operations: ADD_NUMBER, ADD_SCALAR, and MULTIPLY_SCALAR_1
-    // Valid slots (used up to MC 1.20.6): HAND, OFF_HAND, FEET, LEGS, CHEST, HEAD, ANY
-    // Valid slot groups (used on MC 1.20.6+): <@link url https://hub.spigotmc.org/javadocs/spigot/org/bukkit/inventory/EquipmentSlotGroup.html>
-    // Valid attribute names are listed at <@link url https://hub.spigotmc.org/javadocs/spigot/org/bukkit/attribute/Attribute.html>
-    // The default slot/slot group is "any".
-    //
-    // Operation names are based on the Bukkit enum.
-    // ADD_NUMBER corresponds to Mojang "ADDITION" - adds on top of the base value.
-    // ADD_SCALAR corresponds to Mojang "MULTIPLY_BASE" - adds to the total, multiplied by the base value.
-    // MULTIPLY_SCALAR_1 corresponds to Mojang "MULTIPLY_TOTAL", multiplies the final value (after both "add_number" and "add_scaler") by the amount given plus one.
-    //
-    // They are combined like (pseudo-code):
-    // <code>
-    // - define x <[base_value]>
-    // - foreach <all_modifiers[ADD_NUMBER]>:
-    //     - define x:+:<[value]>
-    // - define y <[x]>
-    // - foreach <all_modifiers[ADD_SCALAR]>:
-    //     - define y:+:<[x].mul[<[value]>]>
-    // - foreach <all_modifiers[MULTIPLY_SCALAR_1]>:
-    //     - define y:*:<[value].add[1]>
-    // - determine <[y]>
-    // </code>
-    //
-    // See also <@link url https://minecraft.wiki/w/Attribute#Modifiers>
-    //
-    // For a quick and dirty in-line input, you can do for example: [generic_max_health=<list[<map[key=my_project:add_health;operation=ADD_NUMBER;amount=20;slot=HEAD]>]>]
-    //
-    // For more clean/proper input, instead do something like:
-    // <code>
-    // - definemap attributes:
-    //     generic_max_health:
-    //         1:
-    //             key: my_project:add_health
-    //             operation: ADD_NUMBER
-    //             amount: 20
-    //             slot: head
-    // - inventory adjust slot:head add_attribute_modifiers:<[attributes]>
-    // </code>
-    //
-    // When pre-defining a custom item, instead of this, simply use an item script: <@link language item script containers>. That page shows an example of valid attribute modifiers on an item script.
-    //
-    // -->
 
     public static void register() {
 
@@ -312,7 +151,7 @@ public class EntityAttributeModifiers implements Property {
             try {
                 MapTag input = mechanism.valueAsType(MapTag.class);
                 Attributable ent = getAttributable();
-                ItemAttributeModifiers.parseAttributeModifiers(input, mechanism, attribute -> {
+                AttributeUtil.parseModifiers(input, mechanism, attribute -> {
                     AttributeInstance instance = ent.getAttribute(attribute);
                     if (instance == null) {
                         mechanism.echoError("Attribute " + attribute + " is not applicable to entity of type " + entity.getBukkitEntityType().name());
@@ -347,7 +186,7 @@ public class EntityAttributeModifiers implements Property {
             try {
                 MapTag input = mechanism.valueAsType(MapTag.class);
                 Attributable ent = getAttributable();
-                ItemAttributeModifiers.parseAttributeModifiers(input, mechanism, attribute -> {
+                AttributeUtil.parseModifiers(input, mechanism, attribute -> {
                     AttributeInstance instance = ent.getAttribute(attribute);
                     if (instance == null) {
                         mechanism.echoError("Attribute " + attribute + " is not applicable to entity of type " + entity.getBukkitEntityType().name());
@@ -362,7 +201,7 @@ public class EntityAttributeModifiers implements Property {
                         if (!ex.getMessage().equals("Modifier is already applied on this attribute!")) {
                             throw ex;
                         }
-                        if (ItemAttributeModifiers.MODERN_ATTRIBUTE_FORMAT) {
+                        if (AttributeUtil.MODERN_ATTRIBUTE_FORMAT) {
                             mechanism.echoError("Cannot add attribute with key '" + modifier.getKey() + "' as the entity already has a modifier with the same key.");
                         }
                         else {
@@ -410,7 +249,7 @@ public class EntityAttributeModifiers implements Property {
             for (String toRemove : inputList) {
                 UUID id = null;
                 NamespacedKey key = null;
-                if (ItemAttributeModifiers.MODERN_ATTRIBUTE_FORMAT) {
+                if (AttributeUtil.MODERN_ATTRIBUTE_FORMAT) {
                     key = Utilities.parseNamespacedKey(toRemove);
                 }
                 else {
@@ -422,7 +261,7 @@ public class EntityAttributeModifiers implements Property {
                         continue;
                     }
                     for (AttributeModifier modifier : instance.getModifiers()) {
-                        if (ItemAttributeModifiers.MODERN_ATTRIBUTE_FORMAT ? modifier.getKey().equals(key) : modifier.getUniqueId().equals(id)) {
+                        if (AttributeUtil.MODERN_ATTRIBUTE_FORMAT ? modifier.getKey().equals(key) : modifier.getUniqueId().equals(id)) {
                             instance.removeModifier(modifier);
                             break;
                         }
